remove code duplication for generator driver
Remove the duplicate code for generating the two different templates
