+++ lib/registry.template.ejs
@@ -12,7 +12,7 @@ var authHeader = 'authorization';
  * with multiple running instances of the same REST API.
  *
  */
-var registry = angular.module(<%-: moduleName | q %>, ['ngResource']);
+var module = angular.module(<%-: moduleName | q %>, ['ngResource']);
 
 <% //TODO: docs -%>
 /**
@@ -111,7 +111,7 @@ module.factory(
 <%   }); //action.params.foreach -%>
           },
 <% } -%>
-          url: urlBase + <%-: action.getFullPath() | q %>,
+          url: _urlBase + <%-: action.getFullPath() | q %>,
           method: <%-: action.getHttpMethod() | q %>
         },
 <% }); // meta.methods.foreach -%>
@@ -293,13 +293,13 @@ module
       this.currentUserData= null;
     }
 
-    LoopbackAuth.prototype.switchContext = function(prefix, skipSave) {
+    LoopBackAuth.prototype.switchContext = function(prefix, skipSave) {
       if (!skipSave) {
         this.save();
       }
 
       this.props = propsNames.map(function(name) {
-        return [prefix, name],join('.');
+        return [prefix, name].join('.');
       });
 
       this.props.forEach(function(name) {
@@ -425,7 +425,7 @@ module
      * to the code generator (`lb-ng` or `grunt-loopback-sdk-angular`) is used.
      */
     this.setUrlBase = function(url) {
-      urlBase = url;
+      _urlBase = url;
     };
 
     /**
@@ -437,7 +437,7 @@ module
      * to the code generator (`lb-ng` or `grunt-loopback-sdk-angular`) is used.
      */
     this.getUrlBase = function() {
-      return urlBase;
+      return _urlBase;
     };
 
     this.$get = ['$resource', function($resource) {
