+++ lib/services.js
@@ -5,6 +5,11 @@ ejs.filters.q = function(obj) {
   return JSON.stringify(obj, null, 2 );
 };
 
+module.exports = {
+  services: generateServices,
+  registry: generateRegistry
+};
+
 /**
  * Generate Angular $resource services for the given loopback application.
  *
@@ -24,7 +29,7 @@ ejs.filters.q = function(obj) {
  * @returns {string} The generated javascript code.
  * @header generateServices
  */
-module.exports = function generateServices(app, ngModuleName, apiUrl) {
+function generateServices(app, ngModuleName, apiUrl) {
   ngModuleName = ngModuleName || 'lbServices';
   var ngRegistryName = 'lbServicesRegistry';
   apiUrl = apiUrl || '/';
@@ -41,6 +46,41 @@ module.exports = function generateServices(app, ngModuleName, apiUrl) {
     registryName: ngRegistryName,
     models: models,
     urlBase: apiUrl.replace(/\/+$/, '')
+  });
+};
+
+/**
+ * Generate Angular $resource services for the given loopback application.
+ *
+ * ```js
+ * var generateServices = require('loopback-sdk-angular').services;
+ * var app = require('./server/server');
+ *
+ * var client = generateServices(app, 'lbServices', '/api');
+ * require('fs').writeFileSync('client/loopback.js', client, 'utf-8');
+ * ```
+ *
+ * @param {Object} app The loopback application created via `app = loopback()`.
+ * @param {string=} ngModuleName A name for the generated Angular module.
+ *   Default: `lbServices`.
+ * @param {string=} apiUrl The URL where the client can access the LoopBack
+ *   server app. Default: `/`.
+ * @returns {string} The generated javascript code.
+ * @header generateServices
+ */
+function generateRegistry(app, ngModuleName) {
+  ngModuleName = ngModuleName || 'lbServicesRegistry';
+
+  var models = describeModels(app);
+
+  var registryTemplate = fs.readFileSync(
+    require.resolve('./registry.template.ejs'),
+    { encoding: 'utf-8' }
+  );
+
+  return ejs.render(registryTemplate, {
+    moduleName: ngModuleName,
+    models: models
   });
 };
